+++ lib/livebook_web/live/file_select_component.ex
@@ -22,6 +22,8 @@ defmodule LivebookWeb.FileSelectComponent do
   # To force the component to refetch the displayed files you can
   # `send_update` with `force_reload: true` to the component.
 
+  import LivebookWeb.FileSystemHelpers
+
   alias Livebook.FileSystem
 
   @impl true
